feat : 파티 참여 API를 추가하라

+++ src/main/java/meltingpot/server/domain/repository/PartyRepository.java
@@ -3,5 +3,5 @@ package meltingpot.server.domain.repository;
 import meltingpot.server.domain.entity.party.Party;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PartyRepository extends JpaRepository<Party, Long> {
+public interface PartyRepository extends JpaRepository<Party, Integer> {
 }
